requirements.txt
colorama~=0.4.5
selenium~=4.4.3
cryptography~=37.0.4
python-dotenv~=0.21.0
requests~=2.28.1